+++ app/src/main/java/com/theuhooi/uhooipicbook/data/monsters/MonstersRepository.kt
@@ -1,11 +1,10 @@
 package com.theuhooi.uhooipicbook.data.monsters
 
-import com.theuhooi.uhooipicbook.data.monsters.impl.MonsterItem
+import com.theuhooi.uhooipicbook.data.monsters.impl.MonsterDto
 
 interface MonstersRepository {
     fun loadMonsters(
-        onSuccess: (monsters: List<MonsterItem>) -> Unit,
+        onSuccess: (monsters: List<MonsterDto>) -> Unit,
         onFailure: (error: Throwable) -> Unit
     )
 }
-
